pyproject.toml
[tool.poetry]
name = "diet-calorie-advisor-project"
version = "0.1.0"
description = ""
authors = ["masfana016 <masfanasrullahansari123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-openai = "^0.2.3"
langchain-core = "^0.3.13"
langchain = "^0.3.4"
langchain-community = "^0.3.3"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.1"
beautifulsoup4 = "^4.12.3"
faiss-cpu = "^1.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
